Add macro for adding new chains in core
